refactor : LineSection 객체를 분리
